nominations test

+++ magic512bot/models/nominations.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, String
+from sqlalchemy import BigInteger, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import Base
@@ -7,8 +7,6 @@ from .base import Base
 class Nominations(Base):
     __tablename__ = "nominations"
 
-    id: Mapped[int] = mapped_column(
-        BigInteger, primary_key=True, autoincrement=True
-    )  # Auto-incrementing primary key
-    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)  # Discord user ID
-    format: Mapped[str] = mapped_column(String(55), nullable=False)  # Format nomination
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    format: Mapped[str] = mapped_column(String(55), nullable=False)
